Handle the deprecated message in the invoke method
It was dealt with in the callack. However, the callback of a flow
is called after all the previous commands of the flow, making the
deprecated message appear very late.

Putting the deprecated message in the invocation make the message
appear as soon as the command starts to run, before its flow. This
appears to make much more sense.