Don't try to run the parents once for side effect
Now that we isolated the side effects on the config, using the
config stack, the side effect of running the parent won't be
visible in the children. Hence, the whole path of callbacks must
be run everytime, and the fact we preserve the command line
argument should ensure a coherent behavior.