Don't try to guess the project several times
Instead, let the currentdirectorypreset profile guess the project,
put it into the parameters and let the parse_args loop configure
the whole thing.

This means also that the guess_project method is no longer
supposed to put the value into self._project, which is a
side-effect that made the logic difficult to understand.