Put the complete argument into the parsing context
So that when loading twice the same command, the arguments won't
overlap.

This may happen when for instance you have a command having as
flow the same command with different arguments (using an alias).

Then, the arguments of the flow where put in the command attribute
and the command run with the arguments of the flow dependency.

Because we take extra care to reuse the parsing context, you can
put stuff in it and reuse them afterwards.

So, putting the arguments in the context was much more sensible
and fix the above issue.