Dirty hack to avoid an infinite recursion case
This happens in the case you define an alias inside a group toward
another command of the same group and want to show the help of the
group.

When parsing the arguments of the group, they are saved for later
use. The parsing of --help triggers the help callback. Showing
the help of a group needs to load the sub command (to show their
help messages). When the alias is loaded, it needs to parse the aliased command
to behave like it. This in turn triggers the parsing of the group of
commands. The saved arguments are injected in the new parsing, adding --help and
starting a recursive loop.

This use cass is perfectly normal, as an alias "a.b -> a c" called with "a -a b"
should behave like "a -a c". It would be strange to loose the effect of -a in
the aliased command. Plus, the alias "a.b -> a -a c" called with "a -b b" should
behave as "a -a -b c", remembering for the group the aliased option as well as
the command line one.

In the future, we might want to have a "more elegant" design with only one
parsing of groups of commands, but till then, parsing the group several times is
enough. So for the time being, let's avoid --help to trigger an infinite loop
and call it a day.