feat: Provide a --ask-secret flag that can be used when you don't have a password manager
It asks for the password twice on the command line.

To test it, we use expect to provide interactive input. We needed then to remove
the color so that the output is eventually consistent with what is in the
tutorial page. We did this using the environment variable TERM=dumb to find out
we are not in a interactive session.

Also, we provided the example of calling your own keyring, using the built-in
netrc one.